pyproject.toml
[tool.poetry]
name = "deutschcheckbot"
version = "0.1.0"
description = ""
authors = ["evgen-kulik <e.kulik.atom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.52.0"
python-dotenv = "^1.1.0"
telegram = "^0.0.1"
python-telegram-bot = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
